Delay worker joining backend after dial in
