+++ worker/cmd/worker/worker.go
@@ -1,14 +1,17 @@
 package main
 
 import (
+	"context"
 	"fmt"
-	"github.com/getsentry/sentry-go"
 	"github.com/joschahenningsen/TUM-Live/worker/api"
+	"github.com/joschahenningsen/TUM-Live/worker/cfg"
+	"github.com/joschahenningsen/TUM-Live/worker/pb"
 	"github.com/joschahenningsen/TUM-Live/worker/rest"
 	"github.com/joschahenningsen/TUM-Live/worker/worker"
-	"github.com/makasim/sentryhook"
 	"github.com/pkg/profile"
 	log "github.com/sirupsen/logrus"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials/insecure"
 	"net/http"
 	_ "net/http/pprof"
 	"os"
@@ -20,7 +23,7 @@ import (
 
 // OsSignal contains the current os signal received.
 // Application exits when it's terminating (kill, int, sigusr, term)
-var OsSignal chan os.Signal
+var OsSignal = make(chan os.Signal, 1)
 var VersionTag = "dev"
 
 //prepare checks if the required dependencies are installed
@@ -30,38 +33,52 @@ func prepare() {
 	if err != nil {
 		log.Fatal("ffmpeg is not installed")
 	}
+	_, err = exec.LookPath("tesseract")
+	if err != nil {
+		log.Fatal("tesseract is not installed")
+	}
 }
 
 func main() {
 	prepare()
 
+	// join main tumlive:
+	var conn *grpc.ClientConn
+	var err error
+	// retry connecting to tumlive every 5 seconds until successful
+	for {
+		conn, err = grpc.Dial(fmt.Sprintf("%s:50052", cfg.MainBase), grpc.WithTransportCredentials(insecure.NewCredentials()))
+		if err == nil {
+			break
+		} else {
+			log.Warnf("Could not connect to main tumlive: %v\n", err)
+			time.Sleep(time.Second * 5)
+		}
+	}
+
+	client := pb.NewFromWorkerClient(conn)
+	ctx, cancel := context.WithTimeout(context.Background(), time.Second*5)
+	defer cancel()
+	resp, err := client.JoinWorkers(ctx, &pb.JoinWorkersRequest{
+		Token:    cfg.Token,
+		Hostname: cfg.Hostname,
+	})
+	if err != nil {
+		log.Warnf("Could not join main tumlive: %v\n", err)
+		return
+	}
+	cfg.WorkerID = resp.WorkerId
+	log.Infof("Joined main tumlive with worker id: %s\n", cfg.WorkerID)
 	worker.VersionTag = VersionTag
 	defer profile.Start(profile.MemProfile).Stop()
 	go func() {
 		_ = http.ListenAndServe(":8082", nil) // debug endpoint
 	}()
 
-	OsSignal = make(chan os.Signal, 1)
-
 	// log with time, fmt "23.09.2021 10:00:00"
 	log.SetFormatter(&log.TextFormatter{TimestampFormat: "02.01.2006 15:04:05", FullTimestamp: true})
 	log.SetLevel(log.DebugLevel)
-	if os.Getenv("SentryDSN") != "" {
-		err := sentry.Init(sentry.ClientOptions{
-			Dsn:              os.Getenv("SentryDSN"),
-			TracesSampleRate: 1,
-			Debug:            true,
-			AttachStacktrace: true,
-			Environment:      "Worker",
-		})
-		if err != nil {
-			log.Fatalf("sentry.Init: %s", err)
-		}
-		// Flush buffered events before the program terminates.
-		defer sentry.Flush(2 * time.Second)
-		defer sentry.Recover()
-		log.AddHook(sentryhook.New([]log.Level{log.PanicLevel, log.FatalLevel, log.ErrorLevel, log.WarnLevel}))
-	}
+
 	// setup apis
 	go api.InitApi(":50051")
 	go rest.InitApi(":8060")
